+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 #  CHANGELOG
 
 
-* Feature/Empty_Label_Support/0.1.0
+* Feature/Empty_Label_Support/0.1.0 - 0.1.1
 
 	+ Added `CHANGELOG.md`
+	+ Reuse `TextObserver.OnChangeAction` in `UILabel` text observer extension
 	+ Renamings
 
 * Feature/Empty_Label_Support